+++ src/grib_api_internal.h
@@ -1589,7 +1589,7 @@ typedef struct j2k_encode_helper
 
 } j2k_encode_helper;
 
-#include "grib_api_prototypes.h"
+#include "eccodes_prototypes.h"
 
 
 #ifdef __cplusplus
